+++ packages/core/CHANGELOG.md
@@ -10,9 +10,11 @@ The following is a curated list of changes in the Enact core module, newest chan
 
 ### Changed
 
-- `@core/handle.forwardWithPrevent` to also call `preventDefault()` on the originating event
-
 ### Fixed
+
+## [1.5.0] - 2017-07-19
+
+No significant changes.
 
 ## [1.4.1] - 2017-07-05