+++ packages/spotlight/CHANGELOG.md
@@ -10,9 +10,18 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Changed
 
-- `spotlight` containers to support an array of selectors for `defaultElement`
+- `spotlight` containers using a `restrict` value of 'self-only' will ignore `leaveFor` directives when attempting to leave the container via 5-way
 
 ### Fixed
+
+- `spotlight` to not blur and re-focus an element that is already focused
+
+## [1.5.0] - 2017-07-19
+
+### Changed
+
+- `spotlight` 5-way target selection to ignore empty containers
+- `spotlight` containers to support an array of selectors for `defaultElement`
 
 ## [1.4.1] - 2017-07-05